pyproject.toml
[project]
name = "runpod-qubits"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x>=13.6.0",
    "numpy>=2.3.4",
    "pelicanq-sitecustomize>=0.1.0",
    "runpod>=1.7.13",
    "scipy>=1.16.2",
]
